RFC: Geneve OpenFlow support.
This commit adds support for Geneve flow mods via two nicira extensions.
NXM_NX_TUN_METADATA for being able to match on tun_metadata nxms and
NXAST_RAW_TUN_METADATA for tun_metadata actions.

Matching
========
On the match side multiple NXM_NX_TUN_METADATA nxms can appear in a flow
mod in any order. tun_metadata options are learned as flows get created.
The option space is shared between all the flows for both match and
actions. Since a flow can contain in the worst case all possible options
the max space reserved for the options is limited to 255 bytes.

tun_metadata options are hashed into two tables.

1. maps (class, type) -> (len, offset)

Duplicates are not allowed, this the len is unique to a particular (class,type)

2. maps (offset) -> (class,type,len)

The flow stores the oxm data as it appears (without the length). Given this
and the fact that the metadata in the flow can contain holes the second table
is used to figure out how much to skip to decode the next option.

One of the side effect of maintaining the tables is that commands like
ovs-dpctl that directly talke to datapath do not have access to the table
and cannot use it to dump. Since the dp flow was already created and the
netlink message contains the length of the option this issue can be handled
without needing to consult the table. I plan to implement this via a separate
commit.

Two new commands are added to ovs-appctl to be able to dump and flush the
table.

  tnl/meta/flush
  tnl/meta/show

The other side affect of keeping a table is that when flows are deleted
we cannot necesarily remove and resuse the holes in the option space without
additional complexity. I have not yet figured out how to maintain a reference
count of the flows that are using a particular entry.

ovs-ofctl is expanded to be able to pass multiple tun_metadata options.

example:
ovs-ofctl add-flow br0 "tun_id=0x32,tun_src=12.1.1.5,tun_metadata=1234561122334455667788,tun_metadata=123478aabbccdd,in_port=1 actions=output:2"

Actions
=======
A new action tun_metadata is added. Multiple of these actions can appear in
a flowmod. The actions consult the same table to figure out mappings.

Usage
=====
tun_metadata options are intended to be used for geneve and designed as such.
However they provide an alternate way to configure other types of tunnels.
Consumers of this api need to provide
1. datapath (like geneve kmod)
2. nlattr_to_tun_metadata: function to convert netlink attributes
provided by datapath into tun metadata used for matching.
3. tun_metadata_to_nlattr: function to convert tun metadata into netlink
attributes for the particular datapath.

todo list:
1. add a flag to flow to optimize check for is_all_zeros
2. handle dpctl dump-flows
3. remove the hardcoded if (nxm_class(header) == 1 && nxm_field(header) == 38) {
   in nxm_field_by_header().

Signed-off-by: Madhu Challa <challa@noironetworks.com>